Refactored names and imports Created app_useful_functions file and moved function from variables to it Removed unneeded graphing file
